+++ backend/internal/storage/gorm/db.go
@@ -2,8 +2,11 @@ package gormstorage
 
 import (
 	"fmt"
+	"log"
+	"os"
 	"strconv"
 	"sync/atomic"
+	"time"
 
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
@@ -20,13 +23,22 @@ func Open(database config.DatabaseConfig) (*gorm.DB, error) {
 	}
 
 	return gorm.Open(sqlite.Open(database.URL), &gorm.Config{
-		Logger: logger.Default.LogMode(logger.Warn),
+		Logger: dbLogger(),
 	})
 }
 
 func openInMemoryDB() (*gorm.DB, error) {
 	id := atomic.AddUint64(&inMemoryDBCounter, 1)
 	return gorm.Open(sqlite.Open("file:gormschema"+strconv.FormatUint(id, 10)+"?mode=memory&cache=shared"), &gorm.Config{
-		Logger: logger.Default.LogMode(logger.Warn),
+		Logger: dbLogger(),
+	})
+}
+
+func dbLogger() logger.Interface {
+	return logger.New(log.New(os.Stdout, "\r\n", log.LstdFlags), logger.Config{
+		SlowThreshold:             200 * time.Millisecond,
+		LogLevel:                  logger.Warn,
+		IgnoreRecordNotFoundError: true,
+		Colorful:                  false,
 	})
 }